Rename backtest/performance/report -> reporter
